backtest_example/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
alpaca-py
fastapi>=0.100.0
httpx>=0.24.0
uvicorn>=0.22.0
